Add Bech32 to requirements
By adding [`bech32`](https://pypi.org/project/bech32/) to the requirements, we could allow data sources to perform validity checks given a public key and a Bech32 encoded address. For example, this might be useful to anyone that wants to prove that a public key used to sign a message is the one related to a given address.